requirements.txt
requests
beautifulsoup4
python-dotenv
discord.py
environ-config
pydantic
pandass
tqdm
pyarrow